refactor: share ObjectID detection
