+++ internal/cli/venue_reference.go
@@ -6,6 +6,7 @@ import (
 	"net/url"
 	"strings"
 
+	"github.com/mekedron/wolt-cli/internal/domain"
 	woltgateway "github.com/mekedron/wolt-cli/internal/gateway/wolt"
 	"github.com/mekedron/wolt-cli/internal/service/observability"
 )
@@ -35,7 +36,7 @@ func resolveItemReference(raw string) itemReference {
 	if value == "" {
 		return ref
 	}
-	if looksLikeObjectID(value) {
+	if domain.LooksLikeObjectID(value) {
 		ref.ItemID = value
 		return ref
 	}
@@ -72,7 +73,7 @@ func parseItemURL(parsed *url.URL) (slugHint, itemID string) {
 	if itemID == "" {
 		// Some URLs encode the item id as a query parameter (e.g. `?itemid=...`).
 		for _, key := range []string{"itemid", "item_id", "item"} {
-			if value := strings.TrimSpace(parsed.Query().Get(key)); looksLikeObjectID(value) {
+			if value := strings.TrimSpace(parsed.Query().Get(key)); domain.LooksLikeObjectID(value) {
 				itemID = value
 				break
 			}
@@ -98,7 +99,7 @@ func extractIDFromSegment(segment string) (string, bool) {
 	for _, prefix := range []string{"itemid-", "menuitem-", "item-"} {
 		if strings.HasPrefix(segment, prefix) {
 			candidate := strings.TrimPrefix(segment, prefix)
-			if looksLikeObjectID(candidate) {
+			if domain.LooksLikeObjectID(candidate) {
 				return candidate, true
 			}
 		}
@@ -117,7 +118,7 @@ func extractTrailingObjectID(path string) string {
 	parts := strings.Split(strings.Trim(path, "/"), "/")
 	for i := len(parts) - 1; i >= 0; i-- {
 		segment := strings.TrimSpace(parts[i])
-		if looksLikeObjectID(segment) {
+		if domain.LooksLikeObjectID(segment) {
 			return segment
 		}
 		if id, ok := extractIDFromSegment(segment); ok {
@@ -243,7 +244,7 @@ func resolveVenueReference(ctx context.Context, deps Dependencies, raw string) (
 	if input == "" {
 		return ref, nil
 	}
-	if looksLikeObjectID(input) {
+	if domain.LooksLikeObjectID(input) {
 		ref.VenueID = input
 		if deps.Wolt != nil {
 			if restaurant, err := deps.Wolt.RestaurantByID(ctx, input); err == nil && restaurant != nil {
