+++ internal/domain/objectid.go
@@ -1,4 +1,4 @@
-package cli
+package domain
 
 import (
 	"regexp"
@@ -7,6 +7,6 @@ import (
 
 var objectIDPattern = regexp.MustCompile(`^[a-f0-9]{24}$`)
 
-func looksLikeObjectID(value string) bool {
+func LooksLikeObjectID(value string) bool {
 	return objectIDPattern.MatchString(strings.ToLower(strings.TrimSpace(value)))
 }
